requirements.txt
boto3>=1.34.0
botocore>=1.34.0
PyYAML>=6.0.1
typer>=0.12.3
rich>=13.7.1
tqdm>=4.66.4
transformers>=4.42.0
torch>=2.3.0
accelerate>=0.33.0
orjson>=3.10.0
tenacity>=8.3.0
jsonschema>=4.22.0
shortuuid>=1.0.12
requests>=2.31.0
google-auth>=2.31.0
fastapi>=0.111.0
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1
pydantic>=2.7.0
python-multipart>=0.0.9
aiofiles>=23.2.1

